refactor: extract 'showThemedToast' and 'getDensityAdjustedValue' to :common:android
UIUtils.kt held two unrelated helpers: Toast wrappers and a density scaler.
Both are pure Android utilities with no project deps, so they belong in
:common:android. Splits them into Toast.kt and Density.kt for clarity and
deletes the AnkiDroid-side UIUtils.kt.
