+++ AnkiDroid/src/main/java/com/ichi2/anki/FieldEditLine.kt
@@ -31,6 +31,7 @@ import androidx.annotation.DrawableRes
 import androidx.constraintlayout.widget.ConstraintSet
 import androidx.core.os.ParcelCompat
 import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat
+import com.ichi2.anki.common.utils.android.getDensityAdjustedValue
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.compat.setTooltipTextCompat
 import com.ichi2.anki.databinding.ViewCardMultimediaEditlineBinding
