+++ AnkiDroid/src/main/java/com/ichi2/anki/IntentHandler2.kt
@@ -18,6 +18,7 @@ package com.ichi2.anki
 
 import android.os.Bundle
 import com.ichi2.anki.NoteEditorFragment.Companion.NoteEditorCaller
+import com.ichi2.anki.common.utils.android.showThemedToast
 import com.ichi2.anki.noteeditor.NoteEditorLauncher
 import timber.log.Timber
