+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/OnRenderProcessGoneDelegate.kt
@@ -23,8 +23,8 @@ import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.Lifecycle
 import com.ichi2.anki.AbstractFlashcardViewer
 import com.ichi2.anki.R
+import com.ichi2.anki.common.utils.android.showThemedToast
 import com.ichi2.anki.libanki.CardId
-import com.ichi2.anki.showThemedToast
 import com.ichi2.utils.cancelable
 import com.ichi2.utils.message
 import com.ichi2.utils.positiveButton
