+++ AnkiDroid/src/main/java/com/ichi2/anki/pages/CongratsPage.kt
@@ -39,6 +39,7 @@ import com.ichi2.anki.StudyOptionsActivity
 import com.ichi2.anki.common.time.SECONDS_PER_DAY
 import com.ichi2.anki.common.time.TIME_HOUR
 import com.ichi2.anki.common.time.TIME_MINUTE
+import com.ichi2.anki.common.utils.android.showThemedToast
 import com.ichi2.anki.dialogs.customstudy.CustomStudyDialog
 import com.ichi2.anki.dialogs.customstudy.CustomStudyDialog.CustomStudyAction
 import com.ichi2.anki.launchCatchingIO
@@ -46,7 +47,6 @@ import com.ichi2.anki.launchCatchingTask
 import com.ichi2.anki.observability.ChangeManager
 import com.ichi2.anki.observability.undoableOp
 import com.ichi2.anki.preferences.sharedPrefs
-import com.ichi2.anki.showThemedToast
 import com.ichi2.anki.snackbar.showSnackbar
 import com.ichi2.utils.listItemsAndMessage
 import com.ichi2.utils.negativeButton
