+++ AnkiDroid/src/main/java/com/ichi2/anki/snackbar/Snackbars.kt
@@ -30,7 +30,7 @@ import com.ichi2.anki.BuildConfig
 import com.ichi2.anki.R
 import com.ichi2.anki.common.crashreporting.CrashReportService
 import com.ichi2.anki.common.exception.ManuallyReportedException
-import com.ichi2.anki.showThemedToast
+import com.ichi2.anki.common.utils.android.showThemedToast
 import timber.log.Timber
 
 typealias SnackbarBuilder = Snackbar.() -> Unit
